Adding movies to calendar working. Still need way to delete. Moved Upcoming movies to /movies and renamed Movies to Discover
