--- media-wrangler-front-end/src/components/ComingSoon/ComingSoon.css
@@ -1,189 +1,0 @@
-#movie-search {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 20px;
-  }
-
-h1 {
-    text-align: center;
-    font-size: 36px;
-    margin-bottom: 20px;   
-}
-
-
-.button-container {
-    display: flex;
-    gap: 2px;
-    margin-bottom: 20px;
-    border: 2px solid #ccc;
-    border-radius: 10px;
-    padding: 5px;
-    justify-content: center;
-}
-
-.posterContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 200px;
-    margin: 10px;
-    border-radius: 10px;
-    overflow: hidden;
-    position: relative;
-    transition: transform 0.3s ease;
-  }
-  
-  .posterContainer:hover {
-    transform: scale(1.05);
-  }
-  
-  .posterImage {
-    width: 100%;
-    border-bottom: 1px solid #ccc;
-  }
-
-
-.button {
-    padding: 8px 15px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-    transition: background-color 0.3s;
-    margin: 0;
-}
-
-
-.button.selected {
-    background-color: lightblue;
-}
-
-
-.search-button {
-    padding: 10px 20px;
-    border: none;
-    border-radius: 25px;
-    background-color: #4CAF50;
-    color: white;
-    font-size: 18px;
-    cursor: pointer;
-    transition: background-color 0.3s ease, transform 0.1s ease; 
-    outline: none;
-}
-
-
-.search-button:hover {
-    background-color: #45a049;
-}
-
-
-.search-button:active {
-    background-color: #0a470e;
-    transform: scale(0.98);
-}
-
-
-input[type="text"] {
-    display: block;
-    margin: 0 auto;
-    padding: 10px 20px;
-    font-size: 16px;
-    border: 2px solid #ccc;
-    border-radius: 5px;
-    width: 300px;
-    margin-bottom: 20px;
-}
-
-button {
-    display: block;               
-    margin: 0 auto;               
-    padding: 10px 20px;           
-    font-size: 16px;              
-    border-radius: 5px;           
-    cursor: pointer;             
-    margin-bottom: 20px;          
-}
-
-.movie-item {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 10px;
-  }
-  
-  .releaseDate {
-    font-size: 14px;
-    color: #555;
-    margin-top: 10px;
-    text-align: center;
-  }
-  
-  .releaseDate {
-    font-size: 14px;
-    color: #555;
-    margin-top: 10px;
-    text-align: center;
-  }
-
-  .topRightButton {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    background-color: #4d6a6d;
-    color: white;
-    border: none;
-    border-radius: 12px;
-    width: 40px;
-    height: 40px;
-    cursor: pointer;
-    font-size: 18px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-    transition: background-color 0.3s ease, transform 0.1s ease;
-  }
-  
-  .topRightButton:hover {
-    background-color: #45a049;
-  }
-  
-  .topRightButton:active {
-    background-color: #3e8e41;
-    transform: scale(0.95);
-  }
-
-  .dropdown-container {
-    position: relative;
-  }
-
-  .dropdown-menu {
-    position: absolute;
-    top: 50px;
-    right: 10px;
-    background-color: white;
-    border: 1px solid #ccc;
-    border-radius: 8px;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-    z-index: 1000;
-    padding: 10px;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .dropdown-item {
-    background-color: #4d6a6d;
-    color: white;
-    border: none;
-    padding: 8px 15px;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 14px;
-    text-align: center;
-    margin-bottom: 5px;
-  }
-  
-  .dropdown-item:hover {
-    background-color: #405b5d;
-  }
